mqtt-listen/requirements.txt
paho-mqtt>=2.0.0
pyserial>=3.5
